+++ package.json
@@ -29,7 +29,10 @@
     ]
   },
   "main": "index.js",
-  "devDependencies": {},
+  "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/plugin-transform-private-property-in-object": "^7.22.11"
+  },
   "keywords": [],
   "author": "",
   "license": "ISC",
