+++ src/App.js
@@ -1,19 +1,6 @@
 import React from "react";
-
-const todoList = [
-  {
-    id: "1",
-    title: "Project setup",
-  },
-  {
-    id: "2",
-    title: "Learn first lesson",
-  },
-  {
-    id: "3",
-    title: "Complete assignment",
-  },
-];
+import TodoList from "./TodoList";
+import AddTodoForm from "./AddTodoForm";
 
 const today = new Date();
 function App() {
@@ -23,12 +10,8 @@ function App() {
       <h3>
         Date: {today.getMonth()}/{today.getDate()}/{today.getFullYear()}
       </h3>
-
-      <ul>
-        {todoList.map((listItem) => (
-          <li key={listItem.id}>{listItem.title}</li>
-        ))}
-      </ul>
+      <AddTodoForm />
+      <TodoList />
     </div>
   );
 }
